requirements_v2.txt
catboost==1.2.5
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0.1
requests>=2.31
lightgbm>=4.5.0

tqdm>=4.66
tenacity>=8.2
